refac: Refatoração do servidor para permitir testes
